+++ hub.go
@@ -44,7 +44,7 @@ const (
 	rootUserAgent   = "/golang-event-hubs"
 
 	// Version is the semantic version number
-	Version = "0.3.1"
+	Version = "0.4.0"
 )
 
 type (