+++ changelog.md
@@ -1,7 +1,8 @@
 # Change Log
 
-## `head`
-- add partition key to received event
+## `v0.4.0`
+- add partition key to received event [#43](https://github.com/Azure/azure-event-hubs-go/pull/43)
+- remove `Receive` in eph in favor of `RegisterHandler`, `UnregisterHandler` and `RegisteredHandlerIDs` [#45](https://github.com/Azure/azure-event-hubs-go/pull/45)
 
 ## `v0.3.1`
 - simplify environmental construction by prefering SAS